Fix audio glitch caused by 0-value samples

+++ dynawave.h
@@ -19,7 +19,7 @@ typedef struct AudioState {
 	uint16_t clocks[4];
 	uint16_t periods[4];
 	uint8_t volumes[4];
-	uint8_t out[4];
+	uint16_t out[4];
 	uint16_t lfsr;
 	uint16_t wave_index;
 } AudioState;
